update login and productlist page

+++ src/pages/login/Login.jsx
@@ -1,9 +1,7 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 /* eslint-disable jsx-a11y/alt-text */
 /* eslint-disable jsx-a11y/anchor-is-valid */
-import { useContext , useEffect } from 'react'
-import { NavLink, useHistory } from 'react-router-dom'
-import { UserContext} from '../../context/UserContext'
+import { NavLink } from 'react-router-dom'
 import { useForm } from 'react-hook-form'
 import api from '../../services/authUserAPI'
 import Swal from 'sweetalert2'
@@ -12,47 +10,62 @@ const Login = () => {
 
     document.title = 'Login'
 
-    let history = useHistory()
-
-    // เรียกใช้งาน Context
-    const {user, setUser} = useContext(UserContext)
-    console.log("user", user)
-
     // เรียกใช้งาน React Hook From
     const { register, handleSubmit, errors } = useForm()
 
     // ฟังก์ชันหลังจาก Submit form
     const onSubmit = (data) => {
-        // console.log(data)
-
         const authData = {
-            "identifier": data.email,
+            "identifier":data.email,
             "password": data.password
         }
-
-        // เรียก API Login
+        // console.log(data)
         api.authLogin(authData).then(res => {
-            console.log(res)
+            // console.log(res)
+            // alert('login success')
 
-            // Set Context
-            setUser(res)
+            let timerInterval
+            Swal.fire({
+                html: 'กำลังเข้าสู่ระบบ <b></b>',
+                timer: 2000,
+                timerProgressBar: true,
+                didOpen: () => {
+                    Swal.showLoading()
+                    timerInterval = setInterval(() => {
+                    const content = Swal.getContent()
+                    if (content) {
+                        const b = content.querySelector('b')
+                        if (b) {
+                            b.textContent = Swal.getTimerLeft()
+                        }
+                    }
+                    }, 2000)
+                },
+                willClose: () => {
+                    clearInterval(timerInterval)
+                }
+            }).then((result) => {
+                if (result.dismiss === Swal.DismissReason.timer) {
 
-            // เก็บ Token ลง LocalStorage
-            localStorage.setItem("token", res.data.jwt)
+                    // Set context
+                    localStorage.setItem("token", res.data.jwt)
 
-            // Redirect ไปหน้า dashboad
-            history.push('/dashboard')
+                    // ส่งไปหน้าหลัก
+                    window.location = '/dashboard'
+                }
+            })
 
-        }).catch(error => {
-            // console.log(error)
+        }) 
+        .catch(error => {
+            // console.log(error.response.status)
             if(error.response.status === 400){
                 Swal.fire({
                     icon: 'error',
-                    text: 'ข้อมูลเข้าระบบไม่ถูกต้อง ลองใหม่',
-                    confirmButtonText: 'ปิดหน้าต่าง'
+                    text: 'ข้อมูลเข้าระบบไม่ถูกต้อง ลองใหม่ !',
+                    confirmButtonText:'ปิดหน้าต่าง',
                 })
             }
-        })
+        });
     }
 
     return (
